Sign summary panel - hover highlights all same module instance elements
Implemented some additional hover behaviour in the sign summary window. Now if you hover over an oval or a rectangle module instance button, not only does it get highlighted, but so do ALL of its fellow buttons that refer to the same module instance. This is true not only for the other hand's version of the exact same timing point/interval, but also ALL of the other timing points/intervals that are linked for that same module instance. This makes it clearer to the user which elements in the summary are linked to the same module instance.