save_as bugfix -- indentation does matter!
Previously 'cancel' on the file selection dialog was ignored
and the program saved the corpus. This behaviour took place
because the actual saving function was *outside* the block of
the conditional that checks for a new path.
So SLPAA saved the corpus no matter a new path was
specified or not.

I'm not a semanticist but cancelling out of the file path
dialog does not simply mean 'I don't want a new path'
but it means 'I don't want to save.'
So I indented the lines inside the conditional so that
only when a new path is given should the save function
take place.