rel comp bugfix: Y change from module to artic.
When Y changes from an existing module to an articulator like H2, the existingmodule attribute in the RelationY object does not update. This results in 'ghost' module participating in comparison. For this reason, this commit decides the type of Y (i.e., articulator or existing module) by referring to the boolean attribute existingmodule instead.